feat: edit-products - Now it is possible to edit added products thanks to refactored viewmodel and view which allow for already existing products to be edited. - Further improved the part of the form which is responsible for managing multipliers as now their order is properly maintained. Manipulation of the order is done via drag&drop functionality. - Implemented the process of deletion of a product. - Fixed the drop-down with available multipliers not showing up.

+++ app/src/main/java/com/example/stockmate/data/dao/ProductMultiplierDao.kt
@@ -15,5 +15,5 @@ interface ProductMultiplierDao {
     suspend fun deleteAllMultipliers()
 
     @Query("DELETE FROM product_multipliers WHERE productId IN (:productIds)")
-    suspend fun deleteMultipliersByProductIds(productIds: List<Int>)
+    suspend fun deleteMultipliersByProductIds(productIds: List<Long>)
 }
